+++ src/index.html
@@ -25,7 +25,7 @@
             </button>
         </form>
         <ul class="list-container__items"></ul>
-        <button class="list-container__remove-all">Clear all completed</button>
+        <button type="button" class="list-container__remove-all">Clear all completed</button>
     </div>
 </body>
