+++ src/index.js
@@ -1,28 +1,22 @@
 import './main.scss';
-import { checkList, clearCompletedTask, displayTask, selectATask, spinner } from './modules/controlList';
-import { allTasks as tasks } from './modules/taskManager';
+import { addTask, spinner } from './modules/controlList';
+import { clearCompletedTasks, markTask, selectATask } from './modules/taskManager';
 
-const clearButton = document.querySelector('.list-container__remove-all');
-const checkbox = document.querySelectorAll('input[type=checkbox]');
-const taskName = document.querySelectorAll('.task-name');
-const deleteList = document.querySelectorAll('.bar');
-
-// Load and shows the number of tasks
+// Load and display the number of tasks
 spinner();
 
-// Show Tasks into the dom
-displayTask(tasks);
-
-// Check the task if it's completed or not
-checkList(checkbox);
+// Create and Add new task
+addTask();
 
 /**
- * When the task is selected
- *  > Highlight the list section
- *  > Delete task when the delete icon is pressed
- *  > Update task if the value is changed
+ * 1. Highlight the selected task
+ * 2. Remove the list when the delete icon is pressed
+ * 3. Update task list when the value of the task is changed
  */
-selectATask(taskName, deleteList);
+selectATask();
 
-// Remove completed tasks
-clearCompletedTask(clearButton);
+// Select or Unselect for task status <completed or Not>
+markTask();
+
+// Clear complted tasks
+clearCompletedTasks();
