+++ src/main.scss
@@ -47,6 +47,16 @@ body {
   font-weight: 400;
 }
 
+.icon {
+  color: var(--strike-light);
+  font-size: 1.6rem;
+  cursor: pointer;
+
+  &:hover {
+    color: var(--dark-grey);
+  }
+}
+
 .list-container {
   display: flex;
   flex-direction: column;
@@ -55,6 +65,44 @@ body {
   margin: 10vh auto;
   box-shadow: rgba(17, 17, 26, 0.05) 0 1px 0, rgba(17, 17, 26, 0.1) 0 0 8px;
   border-radius: 0.5rem;
+
+  .checkbox {
+    margin-right: 1rem;
+    width: 1.5rem;
+    height: 1.5rem;
+    position: relative;
+  }
+
+  &__list-content {
+    padding: 0 2rem;
+    border: 1px solid var(--grey);
+    display: flex;
+    align-items: center;
+
+    input,
+    input:focus {
+      border: none;
+      flex: 2;
+      outline: none;
+      padding: 2rem 0;
+      font-family: inherit;
+      font-size: inherit;
+      font-weight: inherit;
+
+      &::placeholder {
+        font-style: italic;
+      }
+    }
+
+    &--btn {
+      border: none;
+      background: none;
+
+      .icon {
+        transform: rotate(90deg);
+      }
+    }
+  }
 
   &__title {
     display: flex;
@@ -74,7 +122,7 @@ body {
       .icon {
         position: relative;
         align-self: flex-end;
-        margin-right: .5rem;
+        margin-right: 0.5rem;
       }
 
       .notification {
@@ -87,7 +135,7 @@ body {
         border-radius: 50%;
         color: var(--white);
         z-index: 5;
-        font-size: .9rem;
+        font-size: 0.9rem;
         text-align: center;
       }
     }
@@ -104,13 +152,6 @@ body {
       display: flex;
       align-items: center;
       cursor: pointer;
-
-      input {
-        margin-right: 1rem;
-        width: 1.5rem;
-        height: 1.5rem;
-        position: relative;
-      }
 
       .input-after::after {
         content: '\f00c';
@@ -141,6 +182,7 @@ body {
 
       .icon-delete {
         position: relative;
+
         &::after {
           content: '\f2ed';
           font: var(--fa-font-solid);
@@ -151,36 +193,6 @@ body {
           background-color: var(--white);
         }
       }
-    }
-  }
-
-  &__list-content {
-    padding: 0 2rem;
-    border: 1px solid var(--grey);
-    display: flex;
-    align-items: center;
-
-    input,
-    input:focus {
-      border: none;
-      flex: 2;
-      outline: none;
-      padding: 2rem 0;
-      font-family: inherit;
-      font-size: inherit;
-      font-weight: inherit;
-
-      &::placeholder {
-        font-style: italic;
-      }
-    }
-
-    &--btn {
-      border: none;
-      background: none;
-      .icon {
-        transform: rotate(90deg);
-      }  
     }
   }
 
@@ -219,16 +231,6 @@ body {
   color: var(--strike);
 }
 
-.icon {
-  color: var(--strike-light);
-  font-size: 1.6rem;
-  cursor: pointer;
-
-  &:hover {
-    color: var(--dark-grey);
-  }
-}
-
 .list-highlight {
   background-color: var(--highlight);
 
@@ -241,6 +243,7 @@ body {
   from {
     transform: rotate(0deg);
   }
+
   to {
     transform: rotate(359deg);
   }
