+++ src/modules/taskManager.js
@@ -1,57 +1,59 @@
-let tasks = [];
+import {
+  removeCompletedTask, removeTask, status, updateTask,
+} from './controlList';
 
-const storage = localStorage.getItem('listItem');
-tasks = storage === null ? [] : JSON.parse(storage);
+const checkbox = document.querySelectorAll('input[type=checkbox]');
+const taskName = document.querySelectorAll('.task-name');
+const deleteList = document.querySelectorAll('.bar');
+const clearButton = document.querySelector('.list-container__remove-all');
 
-const task = {
-  description: '',
-  completed: false,
-  index: 0,
-};
-
-const addTask = (form) => {
-  form.addEventListener('submit', (event) => {
-    task.description = form.elements.list.value;
-    task.index = tasks.length;
-    tasks.push(task);
-    localStorage.setItem('listItem', JSON.stringify(tasks));
-    form.elements.list.value = '';
-    window.location.reload();
-    event.preventDefault();
+const removeHighlight = () => {
+  taskName.forEach((e) => {
+    e.parentElement.classList.remove('list-highlight');
+    e.nextElementSibling.classList.remove('icon-delete');
   });
 };
 
-const updateIndex = () => {
-  for (let i = 0; i < tasks.length; i += 1) {
-    tasks[i].index = i;
+const markTask = () => {
+  for (let i = 0; i < checkbox.length; i += 1) {
+    checkbox[i].addEventListener('change', () => {
+      if (checkbox[i].checked) {
+        checkbox[i].classList.add('input-after');
+        checkbox[i].nextElementSibling.classList.add('checked');
+        status(i, true);
+      } else {
+        checkbox[i].classList.remove('input-after');
+        checkbox[i].nextElementSibling.classList.remove('checked');
+        status(i, false);
+      }
+    });
   }
 };
 
-const removeTask = (index) => {
-  tasks.splice(index, 1);
-  updateIndex();
-  window.localStorage.setItem('listItem', JSON.stringify(tasks));
-};
+const selectATask = () => {
+  for (let i = 0; i < taskName.length; i += 1) {
+    taskName[i].addEventListener('click', () => {
+      removeHighlight();
+      taskName[i].parentElement.classList.add('list-highlight');
+      taskName[i].nextElementSibling.classList.add('icon-delete');
 
+      deleteList[i].addEventListener('click', () => {
+        removeTask(i);
+        window.location.reload();
+      });
 
-const status = (index, type) => {
-  tasks[index].completed = type;
-};
-
-const removeCompletedTask = () => {
-  for (let i = 0; i < tasks.length; i += 1) {
-    if (tasks[i].completed === true) removeTask(i);
+      taskName[i].addEventListener('input', () => {
+        updateTask(i, taskName[i].value);
+      });
+    });
   }
-  updateIndex();
 };
 
-const updateTask = (index, value) => {
-  tasks[index].description = value;
-  window.localStorage.setItem('listItem', JSON.stringify(tasks));
+const clearCompletedTasks = () => {
+  clearButton.addEventListener('click', () => {
+    removeCompletedTask();
+    window.location.reload();
+  });
 };
 
-const allTasks = tasks;
-
-export {
-  addTask, removeTask, updateTask, status, removeCompletedTask, allTasks
-};
+export { markTask, selectATask, clearCompletedTasks };
